Bitcore provider: some cleanup
