+++ src/Client.js
@@ -146,17 +146,6 @@ export default class Client {
   }
 
   /**
-   * Get address UTXOs
-   * @param {!string|string[]|Address|Address[]} addresses - An address or a list of addresses.
-   * @return {Promise<number, InvalidProviderResponseError>} If addresses is given,
-   *  returns the unspent outputs for the addresses provider
-   */
-  async getAddressUtxos (addresses) {
-    const outputs = await this.getMethod('getAddressUtxos')(addresses)
-    return outputs
-  }
-
-  /**
    * Generate a block
    * @param {!number} numberOfBlocks - Number of blocks to be generated
    * @return {Promise<string[], TypeError|InvalidProviderResponseError>} Resolves
@@ -462,7 +451,7 @@ export default class Client {
    * @return {Promise<string>} Resolves with secret
    */
   async generateSecret (message) {
-    const unusedAddress = await this.getMethod('getUnusedAddress')()
+    const unusedAddress = (await this.getMethod('getAddresses')())[0]
     const signedMessage = await this.signMessage(message, unusedAddress)
     const secret = sha256(signedMessage)
     return secret
