+++ src/providers/bitcoin/index.js
@@ -1,4 +1,5 @@
 import BitcoinRPCProvider from './BitcoinRPCProvider'
+import BitcoreRPCProvider from './BitcoreRPCProvider'
 import BitcoinLedgerProvider from './BitcoinLedgerProvider'
 import BitcoinBlockchainAPIProvider from './BitcoinBlockchainAPIProvider'
 import BitcoinSwapProvider from './BitcoinSwapProvider'
@@ -6,6 +7,7 @@ import * as BitcoinUtil from './BitcoinUtil'
 
 export {
   BitcoinRPCProvider,
+  BitcoreRPCProvider,
   BitcoinLedgerProvider,
   BitcoinBlockchainAPIProvider,
   BitcoinSwapProvider,
