--- src/providers/bitcore/index.js
@@ -1,5 +1,0 @@
-import BitcoreRPCProvider from './BitcoreRPCProvider'
-
-export {
-  BitcoreRPCProvider
-}
