+++ src/providers/index.js
@@ -1,9 +1,7 @@
 import * as bitcoin from './bitcoin'
 import * as ethereum from './ethereum'
-import * as bitcore from './bitcore'
 
 export {
   bitcoin,
-  ethereum,
-  bitcore
+  ethereum
 }
